refactor(remote-mobile): scope terminal status recovery to feature
