--- scripts/patches/core/all-linux/webview/remote-terminal-status-recovery/patch.js
@@ -1,21 +1,0 @@
-"use strict";
-
-const {
-  webviewAssetPatch,
-} = require("../../../../descriptor.js");
-const {
-  applyLinuxRemoteTerminalStatusRecoveryPatch,
-} = require("../../../../impl/webview/index.js");
-
-module.exports = [
-  webviewAssetPatch({
-    id: "linux-remote-terminal-status-recovery",
-    phase: "webview-asset",
-    order: 1044,
-    ciPolicy: "optional",
-    pattern: /^app-initial~app-main~.*\.js$/,
-    missingDescription: "app main webview bundle",
-    skipDescription: "Linux remote terminal status recovery patch",
-    apply: applyLinuxRemoteTerminalStatusRecoveryPatch,
-  }),
-];
