+++ scripts/patches/impl/webview/index.js
@@ -869,136 +869,6 @@ function applyLinuxAppServerBackfillWaitPatch(currentSource) {
   return patchedSource;
 }
 
-function applyLinuxRemoteTerminalStatusRecoveryPatch(currentSource) {
-  if (
-    currentSource.includes("codexLinuxRemoteTerminalStatusWaitingOnUserInput") &&
-    currentSource.includes("hasUserInputRequest:codexLinuxRemoteHasUserInputRequest") &&
-    currentSource.includes("&&codexLinuxRemoteHasUserInputRequest")
-  ) {
-    return currentSource;
-  }
-
-  let patchedSource = currentSource;
-  const userInputRequestHelper =
-    "function codexLinuxRemoteHasUserInputRequest(e){try{return Array.isArray(e)&&e.some(e=>e?.method===`item/tool/requestUserInput`||e?.method===`item/tool/requestOptionPicker`||e?.method===`item/tool/requestSetupCodexContextPicker`||e?.method===`item/tool/call`&&(e?.params?.tool===`request_onboarding_input`||e?.params?.tool===`request_option_picker`||e?.params?.tool===`setup_codex_context_picker`||e?.params?.tool===`setup_codex_step`))}catch{return!1}}";
-  const withUserInputHelper = (replacement) =>
-    patchedSource.includes("function codexLinuxRemoteHasUserInputRequest(")
-      ? replacement
-      : `${userInputRequestHelper}${replacement}`;
-  const buildTerminalStatusReplacement = (
-    fnName,
-    sideChatVar,
-    responseProgressVar,
-    systemErrorVar,
-    resumeStateVar,
-    runtimeStatusVar,
-  ) =>
-    `function ${fnName}({hasInProgressSideChat:${sideChatVar},isResponseInProgress:${responseProgressVar},latestTurnHasSystemError:${systemErrorVar},resumeState:${resumeStateVar},threadRuntimeStatus:${runtimeStatusVar},hasUserInputRequest:codexLinuxRemoteHasUserInputRequestPending=!0}){let codexLinuxRemoteTerminalStatusActive=${runtimeStatusVar}?.type===\`active\`,codexLinuxRemoteTerminalStatusActiveFlags=Array.isArray(${runtimeStatusVar}?.activeFlags)?${runtimeStatusVar}.activeFlags:null,codexLinuxRemoteTerminalStatusWaitingOnUserInput=codexLinuxRemoteTerminalStatusActiveFlags?.includes(\`waitingOnUserInput\`)===!0,codexLinuxRemoteTerminalStatusLoading=codexLinuxRemoteTerminalStatusActive&&(${responseProgressVar}===!0||codexLinuxRemoteTerminalStatusActiveFlags==null||codexLinuxRemoteTerminalStatusActiveFlags.length>0&&(!codexLinuxRemoteTerminalStatusWaitingOnUserInput||codexLinuxRemoteHasUserInputRequestPending===!0));return ${sideChatVar}?\`loading\`:${runtimeStatusVar}?.type===\`systemError\`?\`error\`:codexLinuxRemoteTerminalStatusLoading?\`loading\`:${resumeStateVar}===\`needs_resume\`?\`idle\`:${systemErrorVar}?\`error\`:${responseProgressVar}===!0?\`loading\`:\`idle\`}`;
-
-  const terminalStatusPattern =
-    /function ([A-Za-z_$][\w$]*)\(\{hasInProgressSideChat:([A-Za-z_$][\w$]*),isResponseInProgress:([A-Za-z_$][\w$]*),latestTurnHasSystemError:([A-Za-z_$][\w$]*),resumeState:([A-Za-z_$][\w$]*),threadRuntimeStatus:([A-Za-z_$][\w$]*)\}\)\{return \2\?`loading`:\6\?\.type===`systemError`\?`error`:\6\?\.type===`active`\?`loading`:\5===`needs_resume`\?`idle`:\4\?`error`:\3===!0\?`loading`:`idle`\}/u;
-  const oldPatchedTerminalStatusPattern =
-    /function ([A-Za-z_$][\w$]*)\(\{hasInProgressSideChat:([A-Za-z_$][\w$]*),isResponseInProgress:([A-Za-z_$][\w$]*),latestTurnHasSystemError:([A-Za-z_$][\w$]*),resumeState:([A-Za-z_$][\w$]*),threadRuntimeStatus:([A-Za-z_$][\w$]*)\}\)\{let codexLinuxRemoteTerminalStatusActive=\6\?\.type===`active`,codexLinuxRemoteTerminalStatusLoading=codexLinuxRemoteTerminalStatusActive&&\(\3===!0\|\|!Array\.isArray\(\6\.activeFlags\)\|\|\6\.activeFlags\.length>0\);return \2\?`loading`:\6\?\.type===`systemError`\?`error`:codexLinuxRemoteTerminalStatusLoading\?`loading`:\5===`needs_resume`\?`idle`:\4\?`error`:\3===!0\?`loading`:`idle`\}/u;
-
-  let terminalStatusFnName = null;
-
-  if (terminalStatusPattern.test(patchedSource)) {
-    patchedSource = patchedSource.replace(
-      terminalStatusPattern,
-      (_match, fnName, sideChatVar, responseProgressVar, systemErrorVar, resumeStateVar, runtimeStatusVar) => {
-        terminalStatusFnName = fnName;
-        return withUserInputHelper(
-          buildTerminalStatusReplacement(
-            fnName,
-            sideChatVar,
-            responseProgressVar,
-            systemErrorVar,
-            resumeStateVar,
-            runtimeStatusVar,
-          ),
-        );
-      },
-    );
-  } else if (oldPatchedTerminalStatusPattern.test(patchedSource)) {
-    patchedSource = patchedSource.replace(
-      oldPatchedTerminalStatusPattern,
-      (_match, fnName, sideChatVar, responseProgressVar, systemErrorVar, resumeStateVar, runtimeStatusVar) => {
-        terminalStatusFnName = fnName;
-        return withUserInputHelper(
-          buildTerminalStatusReplacement(
-            fnName,
-            sideChatVar,
-            responseProgressVar,
-            systemErrorVar,
-            resumeStateVar,
-            runtimeStatusVar,
-          ),
-        );
-      },
-    );
-  }
-
-  const pendingRequestPattern =
-    /function ([A-Za-z_$][\w$]*)\(\{pendingRequestType:([A-Za-z_$][\w$]*),requests:([A-Za-z_$][\w$]*),resumeState:([A-Za-z_$][\w$]*),threadRuntimeStatus:([A-Za-z_$][\w$]*)\}\)\{return \3==null\|\|\4==null\?null:\4===`needs_resume`\?\5\?\.type===`active`&&\5\.activeFlags\.includes\(`waitingOnApproval`\)&&([A-Za-z_$][\w$]*)\(\3\)\?`approval`:\5\?\.type===`active`&&\5\.activeFlags\.includes\(`waitingOnUserInput`\)\?`response`:null:([A-Za-z_$][\w$]*)\(\2\)\?`approval`:\2===`userInput`\?`response`:null\}/u;
-  let pendingRequestFnName = null;
-  if (pendingRequestPattern.test(patchedSource)) {
-    patchedSource = patchedSource.replace(
-      pendingRequestPattern,
-      (_match, fnName, pendingTypeVar, requestsVar, resumeStateVar, runtimeStatusVar, approvalRequestFn, approvalTypeFn) => {
-        pendingRequestFnName = fnName;
-        return withUserInputHelper(
-          `function ${fnName}({pendingRequestType:${pendingTypeVar},requests:${requestsVar},resumeState:${resumeStateVar},threadRuntimeStatus:${runtimeStatusVar}}){return ${requestsVar}==null||${resumeStateVar}==null?null:${resumeStateVar}===\`needs_resume\`?${runtimeStatusVar}?.type===\`active\`&&Array.isArray(${runtimeStatusVar}?.activeFlags)&&${runtimeStatusVar}.activeFlags.includes(\`waitingOnApproval\`)&&${approvalRequestFn}(${requestsVar})?\`approval\`:${runtimeStatusVar}?.type===\`active\`&&Array.isArray(${runtimeStatusVar}?.activeFlags)&&${runtimeStatusVar}.activeFlags.includes(\`waitingOnUserInput\`)&&codexLinuxRemoteHasUserInputRequest(${requestsVar})?\`response\`:null:${approvalTypeFn}(${pendingTypeVar})?\`approval\`:${pendingTypeVar}===\`userInput\`?\`response\`:null}`,
-        );
-      },
-    );
-  } else {
-    const existingPendingRequestPattern =
-      /function ([A-Za-z_$][\w$]*)\(\{pendingRequestType:[A-Za-z_$][\w$]*,requests:[A-Za-z_$][\w$]*,resumeState:[A-Za-z_$][\w$]*,threadRuntimeStatus:[A-Za-z_$][\w$]*\}\)\{[^}]*codexLinuxRemoteHasUserInputRequest/u;
-    const match = patchedSource.match(existingPendingRequestPattern);
-    pendingRequestFnName = match?.[1] ?? null;
-  }
-
-  if (terminalStatusFnName != null && pendingRequestFnName != null) {
-    const pendingCallPattern = new RegExp(
-      `${pendingRequestFnName}\\(\\{pendingRequestType:[^{}]+?,requests:([^{}]*\\([^{}]*\\)[^{}]*?),resumeState:[^{}]+?,threadRuntimeStatus:[^{}]+?\\}\\)`,
-      "u",
-    );
-    const pendingCallMatch = patchedSource.match(pendingCallPattern);
-    const requestExpression = pendingCallMatch?.[1] ?? null;
-    const terminalCallPattern = new RegExp(
-      `${terminalStatusFnName}\\(\\{hasInProgressSideChat:([^{}]+?),isResponseInProgress:([^{}]+?),resumeState:([^{}]+?),threadRuntimeStatus:([^{}]+?),latestTurnHasSystemError:([^{}]+?)\\}\\)`,
-      "u",
-    );
-    if (requestExpression != null && terminalCallPattern.test(patchedSource)) {
-      patchedSource = patchedSource.replace(
-        terminalCallPattern,
-        `${terminalStatusFnName}({hasInProgressSideChat:$1,isResponseInProgress:$2,resumeState:$3,threadRuntimeStatus:$4,latestTurnHasSystemError:$5,hasUserInputRequest:codexLinuxRemoteHasUserInputRequest(${requestExpression})})`,
-      );
-    } else if (
-      patchedSource.includes("pendingRequestType") &&
-      patchedSource.includes("hasInProgressSideChat") &&
-      !patchedSource.includes("hasUserInputRequest:codexLinuxRemoteHasUserInputRequest")
-    ) {
-      console.warn(
-        "WARN: Could not wire remote terminal status to pending user-input requests — stale waiting-user-input recovery may be incomplete",
-      );
-    }
-  }
-
-  if (
-    currentSource.includes("hasInProgressSideChat") &&
-    currentSource.includes("isResponseInProgress") &&
-    currentSource.includes("threadRuntimeStatus") &&
-    patchedSource === currentSource
-  ) {
-    console.warn(
-      "WARN: Could not find remote terminal status insertion point — skipping Linux remote terminal status recovery patch",
-    );
-  }
-
-  return patchedSource;
-}
-
 function applyLinuxI18nGatePatch(currentSource) {
   const alreadyPatchedI18nGateRegexes = [
     /([A-Za-z_$][\w$]*)=[^;]*?\.get\(`enable_i18n`,!1\)[^;]*;let [^;]*,([A-Za-z_$][\w$]*)=[A-Za-z_$][\w$]*\([A-Za-z_$][\w$]*\.localeOverride\),[A-Za-z_$][\w$]*=\1\|\|\2!=null/u,
@@ -1972,7 +1842,6 @@ function patchCommentPreloadBundle(extractedDir) {
 module.exports = {
   applyBrowserAnnotationScreenshotPatch,
   applyLinuxAppServerBackfillWaitPatch,
-  applyLinuxRemoteTerminalStatusRecoveryPatch,
   applyLinuxAppServerFeatureEnablementPatch,
   applyAutomationUpdateEagerToolPatch,
   applyLinuxChatSearchHydrationPatch,
